Collision detection bug fixed
The rigidbody froze the ship in the z axis, but the player controller moved the triggers to z=0 (even though this wasn't shown in the inspector) so it looked like the ship was hitting objects before they got to the ship. When objects would hit z=0, even though it looked like the ship was at z = -34, it would hit the ships triggers at z=0.